New Version
Now we have an auto mod updater for the server, and a modpack version checker for single player.

Added these so uploading mods is no longer required, leaving room for shaders and other files for the server.

Dont forget to check the changelog, i recommend some Runtime arguments to make the game run 20% ish smoother (with shaders on med with i7-9700)